SW-85. Fix in FW.

+++ firmware/openos/openwsn/02b-MAChigh/res.c
@@ -44,9 +44,9 @@ status information about several modules in the OpenWSN stack.
 \returns TRUE if this function printed something, FALSE otherwise.
 */
 bool debugPrint_myDAGrank() {
-   uint8_t output=0;
+   uint16_t output=0;
    output = neighbors_getMyDAGrank();
-   openserial_printStatus(STATUS_DAGRANK,(uint8_t*)&output,sizeof(uint8_t));
+   openserial_printStatus(STATUS_DAGRANK,(uint16_t*)&output,sizeof(uint16_t));
    return TRUE;
 }
